requirements.txt
black==23.7.0
cloudinary==1.34.0
django-cors-headers==4.2.0
django==4.2.4
dj_database_url
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
dj-rest-auth==4.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
isort==5.12.0
pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
whitenoise==6.5.0

